Finite sample/ large sample Likelihood ratio test for circadian pattern detection

Test the significance of circadian curve fitting using finite sample likelihood ratio test

Usage

rhytmicity_test(
  tt,
  yy,
  parStart = list(amp = 3, phase = 0, offset = 0, period = 12),
  FN = TRUE
)

Arguments

tt

Time vector / scanner rotation position

yy

Value

parStart

Named list with starting values for parameters: `amp`, `phase`, `offset`, and `period`.

FN

Type of test to use, TRUE = "FN" (finite) for 'F-like Test or FALSE = "LS" (large samples) for Likilihood Ratio Test. Default is finite sample.

Value

A list of amp, phase, offset, sigma02, sigmaA2, l0, l1, df, stat, and pvalue. model: \(y= A * sin(2*pi*x+B)+C\)

y

a 1*n vector of data y

A

estimated A^hat from fit_sine_curve

B

estimated B^hat from fit_sine_curve

C

estimated C^hat from fit_sine_curve

sigma0

sigma0^hat under H0

sigmaA

sigmaA^hat under H1

n

length of data y

df0

df under H0

df1

df under H1

amp

Amplitude based on formula 1

phase

Phase based on formula 1, phase is restricted within (0, period)

peakTime

Phase based on formula 1, peakTime is restricted within (0, period). phase + peakTime = period/4

offset

Basal level(vertical shift) based on formula 1 or on formula 2

sigma02

Variance estimate under the null (intercept only)

sigmaA2

Variance estimate under the alternative (since curve fitting)

l0

Log likelihood under the null (intercept only)

l1

Log likelihood under the alternative (since curve fitting)

df

Degree of freedom for the LR test

stat

LR statistics

pvalue

P-value from the LR test

R2

Pseudo R2 defined as (tss - rss)/tss

Details

Formula 1: \(yy = amp \times sin(2\pi/period \times (phase + tt)) + offset\) Formula 2: \(yy = A \times sin(2\pi/period \times tt) + B * cos(2*pi/period * tt) + offset\)

sourced from: https://rdrr.io/github/diffCircadian/diffCircadian/ under GPL3 license author: Zhiguang Huo, Haocheng Ding Feb. 24, 2023, 9:07 a.m.; Version 0.0.0 journal article: https://doi.org/10.1093/bib/bbab224 tutorial (for circadian data): http://htmlpreview.github.io/?https://github.com/diffCircadian/diffCircadian/blob/master/vignettes/diffCircadian_tutorial.html

Author

Caleb (Zhiguang Huo)

Examples

set.seed(32608)
n <- 50
Period <- 24
tt <- runif(n,0,Period)
Amp <- 2
Phase <- 6
Offset <- 3
yy <- Amp * sin(2*pi/Period * (tt + Phase)) + Offset + rnorm(n,0,1)
rhytmicity_test(tt, yy,  FN =TRUE, parStart = list(amp = 3, phase = 0, offset = 0, period = 20) )
#> $amp
#> [1] 2.367137
#> 
#> $phase
#> [1] 11.98315
#> 
#> $offset
#> [1] 3.645654
#> 
#> $period
#> [1] 31.09577
#> 
#> $peakTime
#> [1] 26.88657
#> 
#> $sigma02
#> [1] 2.500877
#> 
#> $sigmaA2
#> [1] 0.5191673
#> 
#> $l0
#> [1] -93.86296
#> 
#> $l1
#> [1] -54.5587
#> 
#> $df
#>   df1 df2
#> 1   2  47
#> 
#> $stat
#>     F_stat
#> 1 89.70169
#> 
#> $pvalue
#> [1] 9.00641e-17
#> 
#> $R2
#> [1] 0.7924059
#>