Stitch Sequential Images

This function stitches a sequence of images together by aligning overlapping regions based on cross-correlation or phase-correlation.

Usage

stitch_sequential_images(
  imgs,
  overlap_px = 200,
  side1 = "bottom",
  method = "crosscorr"
)

Arguments

imgs

A character vector of file paths to the images to be stitched.

overlap_px

An integer specifying the number of pixels to use to search for overlap between images. Default is 200. The CI-600 scanner produces 100-200 px overlap at 300dpi.

side1

A string specifying the side of the previous image used for overlap (e.g., "bottom", "top", "left", "right"). Default is "bottom".

method

A string specifying the method for alignment, either "crosscorr" or "phasecorr". Default is "crosscorr".

Value

A stitched image as a cimg object.

Examples

if (FALSE) { # \dontrun{
imgs <- c("image1.jpg", "image2.jpg", "image3.jpg")
stitched <- stitch_sequential_images(imgs = imgs,
                    overlap_px = 250, side1 = "bottom")
} # }